+++ pmd-core/src/test/java/net/sourceforge/pmd/cache/FileAnalysisCacheTest.java
@@ -8,6 +8,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -215,6 +216,22 @@ public class FileAnalysisCacheTest {
                 reloadedCache.isUpToDate(sourceFile));
     }
 
+    @Test
+    public void testClasspathNonExistingEntryIsIgnored() throws MalformedURLException, IOException {
+        final RuleSets rs = mock(RuleSets.class);
+        final ClassLoader cl = mock(ClassLoader.class);
+        
+        System.setProperty("java.class.path", System.getProperty("java.class.path") + File.pathSeparator
+                + tempFolder.getRoot().getAbsolutePath() + File.separator + "non-existing-dir");
+        
+        final FileAnalysisCache reloadedCache = new FileAnalysisCache(newCacheFile);
+        try {
+            reloadedCache.checkValidity(rs, cl);
+        } catch (final Exception e) {
+            fail("Validity check failed when classpath includes non-existing directories");
+        }
+    }
+    
     @Test
     public void testClasspathChangeInvalidatesCache() throws MalformedURLException, IOException {
         final RuleSets rs = mock(RuleSets.class);